refactor(cli): package split exemplar + shared versioned spec loader
Two remaining seocho-vh0 deliverables, both riding the guards PR #480 laid:

cli.py becomes the seocho/cli package, with the legacy body as __init__.py
so 'seocho.cli:main', 'from seocho.cli import ...' and monkeypatched
attributes (seocho.cli.Seocho, seocho.cli.serve_local_runtime in
test_sdk_cli) keep resolving against real module globals. The ontology
group — the 280-line worst offender — is the first migration:
seocho/cli/ontology.py holds its parser and handler together and plugs in
through register_group; its LOCAL_COMMANDS entry, dispatch branch, parser
block and _cmd_ontology are gone from the legacy body. Remaining groups
follow this file's shape one at a time. The parser contract tests pass
unchanged — same tree, different wiring — and the extraction surfaced one
real dependency (Path was a module global the handler silently leaned on),
caught by the existing ontology CLI tests.

seocho/spec_loader.py is the shared home for the generic spec machinery
(SpecError error-list exception, env interpolation, unknown-key
did-you-mean, schema_version checking). run_spec.py now consumes it —
RunSpecError subclasses SpecError, existing except-clauses hold — and run
specs accept an optional 'schema_version: 1', where an unsupported version
is a printed migration signal instead of a misparse. seocho-policy.yaml
starts from this loader instead of copying one of four ancestors.

Also found while here: extraction/tests/test_sdk_cli.py imports the
nonexistent seocho.types and has been uncollectable on main (not in CI) —
ticketed as seocho-7a2 rather than drive-by-fixed.

bd: seocho-vh0 (completes the ticket)
Validation: bash scripts/ci/run_basic_ci.sh (679 passed, 3 skipped;
parser-contract + ontology CLI + run-spec suites green); ruff clean.
